requirements.txt
# Core FastAPI
fastapi>=0.108.0
uvicorn>=0.25.0
gunicorn>=20.1.0
pydantic>=2.0.0
python-multipart
python-dotenv

# Existing dependencies
serpapi>=0.1.5
openai>=1.51.0
firebase-admin>=6.5.0
requests>=2.32.3
payos>=1.0.0

# Vector Search - Semantic Search System
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utilities
scikit-learn>=1.3.0

